requirements.txt
yfinance==0.2.54
pandas==2.2.1
google-cloud-storage==2.14.0
python-dotenv==1.0.1 
requests==2.31.0